requirement.txt
Brotli==1.0.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dash==1.13.4
dash-core-components==1.10.1
dash-html-components==1.0.3
dash-renderer==1.5.1
dash-table==4.8.1
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.2
nltk==3.5
numpy==1.19.0
oauthlib==3.1.0
pandas==1.0.5
Pillow==7.2.0
plotly==4.8.2
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.15.0
textblob==0.15.3
tqdm==4.47.0
tweepy==3.8.0
urllib3==1.25.9
Werkzeug==1.0.1
wordcloud==1.7.0
